requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
python-dotenv
ipykernel
Ipython
arxiv
wikipedia